@@ -19,26 +19,26 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@discordjs/uws": "11.149.1",
-    "better-sqlite3": "5.4.0",
-    "bufferutil": "4.0.1",
-    "discord.js": "11.5.0",
-    "erlpack": "github:discordapp/erlpack",
-    "googleapis": "40.0.0",
-    "libsodium-wrappers": "0.7.4",
-    "miniget": "1.5.3",
-    "moment": "2.24.0",
-    "node-opus": "0.3.2",
-    "ytdl-core": "0.29.3"
+    "better-sqlite3": "^5.4.3",
+    "bufferutil": "^4.0.1",
+    "discord.js": "^11.5.1",
+    "erlpack": "github:hammerandchisel/erlpack",
+    "googleapis": "^47.0.0",
+    "libsodium-wrappers": "^0.7.6",
+    "miniget": "^1.6.1",
+    "moment": "^2.24.0",
+    "node-opus": "^0.3.3",
+    "opusscript": "0.0.7",
+    "ytdl-core": "^1.0.6"
   },
   "devDependencies": {
-    "@types/better-sqlite3": "5.4.0",
-    "@types/node": "12.0.2",
-    "nodemon": "1.19.0",
-    "prettier": "1.17.1",
-    "tslint": "5.16.0",
-    "tslint-config-prettier": "1.18.0",
-    "tslint-eslint-rules": "5.4.0",
-    "typescript": "3.4.5"
+    "@types/better-sqlite3": "^5.4.0",
+    "@types/node": "^13.1.7",
+    "nodemon": "^2.0.2",
+    "prettier": "^1.19.1",
+    "tslint": "^5.20.1",
+    "tslint-config-prettier": "^1.18.0",
+    "tslint-eslint-rules": "^5.4.0",
+    "typescript": "^3.7.4"
   }
 }
